fix: fixed int to Long/long and renamed package
