+++ src/main/java/com/learn/gt/user/service/DemoApplication.java
@@ -1,4 +1,4 @@
-package com.user_service;
+package com.learn.gt.user.service;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
