+++ src/main/java/com/learn/gt/user/service/controller/UsersDataController.java
@@ -1,10 +1,10 @@
-package com.user_service.controller;
+package com.learn.gt.user.service.controller;
 
 
-import com.user_service.dtos.IdDto;
-import com.user_service.dtos.UserDto;
-import com.user_service.model.User;
-import com.user_service.service.UserServiceImpl;
+import com.learn.gt.user.service.dtos.IdDto;
+import com.learn.gt.user.service.dtos.UserDto;
+import com.learn.gt.user.service.model.User;
+import com.learn.gt.user.service.service.UserServiceImpl;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -21,13 +21,13 @@ public class UsersDataController {
         this.userServiceImpl = userServiceImpl;
     }
 
-    @PostMapping()
+    @PostMapping
     public ResponseEntity<IdDto> createUserEntry(@RequestBody @Valid UserDto userDto) {
        return ResponseEntity.ok(new IdDto(userServiceImpl.registerUser(userDto)));
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<User> findUserEntry(@PathVariable("id") Integer id) {
+    public ResponseEntity<User> findUserEntry(@PathVariable("id") Long id) {
         return ResponseEntity.ok(userServiceImpl.findUser(id));
     }
