+++ src/main/java/com/learn/gt/user/service/dtos/CreateUserDto.java
@@ -1,0 +1,4 @@
+package com.learn.gt.user.service.dtos;
+
+public class CreateUserDto {
+}
