+++ src/main/java/com/learn/gt/user/service/dtos/UserDto.java
@@ -1,16 +1,14 @@
-package com.user_service.dtos;
+package com.learn.gt.user.service.dtos;
 
 
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.RequiredArgsConstructor;
 
 @Data
 @NoArgsConstructor
@@ -19,7 +17,7 @@ public class UserDto {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
 
     @NotNull(message = "username is required")
     @Size(min = 1, max = 100)
