+++ src/main/java/com/learn/gt/user/service/exceptions/UserAlreadyExists.java
@@ -1,4 +1,4 @@
-package com.user_service.exceptions;
+package com.learn.gt.user.service.exceptions;
 
 public class UserAlreadyExists extends RuntimeException {
     public UserAlreadyExists(String message){
