+++ src/main/java/com/learn/gt/user/service/repository/UserRepository.java
@@ -1,13 +1,13 @@
-package com.user_service.repository;
+package com.learn.gt.user.service.repository;
 
-import com.user_service.model.User;
+import com.learn.gt.user.service.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     public Optional<User> findByEmail(String email);
 
 }
