+++ src/main/java/com/learn/gt/user/service/service/UserService.java
@@ -1,0 +1,9 @@
+package com.learn.gt.user.service.service;
+
+import com.learn.gt.user.service.dtos.UserDto;
+import com.learn.gt.user.service.model.User;
+
+public interface UserService {
+    Long registerUser(UserDto userDto);
+    User findUser(Long id);
+}
