+++ src/main/java/com/learn/gt/user/service/service/mapper/Mapper.java
@@ -1,4 +1,4 @@
-package com.user_service.service.mapper;
+package com.learn.gt.user.service.service.mapper;
 
 public interface Mapper <F, T> {
     T map(F from);
