--- src/main/java/com/user_service/dtos/CreateUserDto.java
@@ -1,4 +1,0 @@
-package com.user_service.dtos;
-
-public class CreateUserDto {
-}
