+++ src/test/java/com/learn/gt/user/service/DemoApplicationTests.java
@@ -1,4 +1,4 @@
-package com.user_service;
+package com.learn.gt.user.service;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
